Updated .gitignore to include all .ini files. Moved all functions related to configparser to the config_parser.py

+++ config_creator.py
@@ -1,24 +1,30 @@
 """Helper module for creating the `config.ini` configuration file."""
 import configparser
+import os
 
 
-def create_config_file():
+def check_for_config_file():
     """Creates a default version of the `config.ini` file in the root directory."""
-    config = configparser.ConfigParser()
-    config["DEFAULT"] = {
-        "embed_channel_id": "False",
-        "embed_message_id": "False",
-        "embed_description": "False",
-    }
+    config_file_exists = os.path.exists("config.ini")
+    if not config_file_exists:
+        config = configparser.ConfigParser()
+        config["DEFAULT"] = {
+            "embed_channel_id": "False",
+            "embed_message_id": "False",
+            "embed_description": "False",
+        }
 
-    config["GlobalVariables"] = {
-        "embed_channel_id": "False",
-        "embed_message_id": "False",
-        "embed_description": "False",
-    }
+        config["GlobalVariables"] = {
+            "embed_channel_id": "False",
+            "embed_message_id": "False",
+            "embed_description": "False",
+        }
 
-    with open("config.ini", "w", encoding="utf-8") as configfile:
-        config.write(configfile)
+        with open("config.ini", "w", encoding="utf-8") as configfile:
+            config.write(configfile)
+        print("\nCreated new config.ini file.\n")
+    else:
+        print("\nFound exisisting config file.\n")
 
 
 def save_to_config(**variables):
